pyproject.toml
[tool.poetry]
name = "neighborhood_evaluation"
version = "0.1.0"
description = "An innovative app for evaluating and recommending neighborhoods based on user preferences"
authors = ["Daniil Strizhov"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12.5"
crewai = {version = "0.30.8", extras = ["tools"]}
langchain = ">=0.1.0"
langchain-core = "0.1.52"
langchain-groq = "0.1.4"
unstructured = "0.10.25"
pyowm = "3.3.0"
python-dotenv = "1.0.0"
langchain_openai = "0.0.5"
embedchain = "0.1.98"

[tool.pyright]
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
select = ["E", "W", "F", "I", "B", "C4", "ARG", "SIM"]
ignore = ["W291", "W292", "W293"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"